feat: refactor layout and enhance SEO components for improved user experience
Reorganized layout components to streamline metadata management and improve SEO. Introduced new components for structured data and SEO content, enhancing search engine visibility. Updated the theme toggle functionality and refined the GitHub contributions display logic for better performance and user interaction. Removed deprecated code and comments for cleaner implementation.
