+++ components/contribution-grid.tsx
@@ -28,35 +28,30 @@ const getContributionColor = (level: number) => {
 
 export const ContributionGrid = ({ weeks, monthLabels }: ContributionGridProps) => {
   return (
-    <div className="overflow-x-auto scrollbar-none scroll-right-initial">
-      <div className="min-w-max scroll-right-content">
-        {/* Month labels aligned with week columns */}
-        <div className="flex items-start gap-1 text-xs text-muted-foreground mb-2 animate-fade-in" style={{ animationDelay: '0.1s', animationFillMode: 'both' }}>
-          <div className="w-8 flex-shrink-0" />
-          <div className="flex gap-1">
-            {monthLabels.map((label, i) => (
-              <div key={`ml-${i}`} className="w-3 text-center flex-shrink-0">
-                {label}
-              </div>
-            ))}
-          </div>
-        </div>
+    <div className="flex gap-1 animate-fade-in" style={{ animationDelay: '0.1s', animationFillMode: 'both' }}>
+      <div className="w-8 flex flex-col gap-1 text-xs text-muted-foreground flex-shrink-0">
+        <div className="h-3"></div>
+        <div className="h-3">Mon</div>
+        <div className="h-3"></div>
+        <div className="h-3">Wed</div>
+        <div className="h-3"></div>
+        <div className="h-3">Fri</div>
+        <div className="h-3"></div>
+      </div>
 
-        {/* Contribution grid */}
-        <div className="flex gap-1 animate-fade-in" style={{ animationDelay: '0.2s', animationFillMode: 'both' }}>
-          {/* Day labels */}
-          <div className="w-8 flex flex-col gap-1 text-xs text-muted-foreground flex-shrink-0">
-            <div className="h-3"></div>
-            <div className="h-3">Mon</div>
-            <div className="h-3"></div>
-            <div className="h-3">Wed</div>
-            <div className="h-3"></div>
-            <div className="h-3">Fri</div>
-            <div className="h-3"></div>
+      <div className="overflow-x-auto scrollbar-none scroll-right-initial">
+        <div className="min-w-max scroll-right-content">
+          <div className="flex items-start gap-1 text-xs text-muted-foreground mb-2 animate-fade-in" style={{ animationDelay: '0.2s', animationFillMode: 'both' }}>
+            <div className="flex gap-1">
+              {monthLabels.map((label, i) => (
+                <div key={`ml-${i}`} className="w-3 text-center flex-shrink-0">
+                  {label}
+                </div>
+              ))}
+            </div>
           </div>
 
-          {/* Grid */}
-          <div className="flex gap-1 min-w-max">
+          <div className="flex gap-1 min-w-max animate-fade-in" style={{ animationDelay: '0.3s', animationFillMode: 'both' }}>
             {weeks.map((week, weekIndex) => (
               <div key={weekIndex + 1} className={`flex flex-col gap-1 flex-shrink-0`}>
                 {week.map((day: ContributionDay | null, dayIndex: number) => (
@@ -65,7 +60,7 @@ export const ContributionGrid = ({ weeks, monthLabels }: ContributionGridProps) 
                     className={`w-3 h-3 rounded-sm transition-all duration-300 hover:scale-110 animate-fade-in ${day ? getContributionColor(day.level) : "bg-transparent"}`}
                     title={day ? `${day.count} contributions on ${day.date}` : ""}
                     style={{ 
-                      animationDelay: `${0.3 + (weekIndex * 7 + dayIndex) * 0.005}s`,
+                      animationDelay: `${0.4 + (weekIndex * 7 + dayIndex) * 0.005}s`,
                       animationFillMode: 'both'
                     }}
                   />
