+++ components/seo-content.tsx
@@ -1,0 +1,38 @@
+export const SEOContent = () => {
+  return (
+    <div className="sr-only" aria-hidden="true">
+      <h1>Ivan Maier Gallardo - Frontend Engineer Portfolio</h1>
+      <p>
+        Welcome to the personal portfolio of Ivan Maier Gallardo, a skilled Frontend Engineer 
+        with nearly 4 years of experience in building scalable web applications. Ivan Maier Gallardo 
+        specializes in React, TypeScript, and modern web technologies, currently working at 
+        MercadoLibre in Buenos Aires, Argentina.
+      </p>
+      <h2>About Ivan Maier Gallardo</h2>
+      <p>
+        Ivan Maier Gallardo is a passionate Frontend Engineer who has dedicated his career to 
+        creating exceptional user experiences through innovative web technologies. With expertise 
+        in React, TypeScript, JavaScript, and Next.js, Ivan Maier Gallardo has contributed to 
+        numerous high-impact projects at MercadoLibre.
+      </p>
+      <h2>Professional Experience</h2>
+      <p>
+        Ivan Maier Gallardo currently serves as a Frontend Engineer at MercadoLibre, where he 
+        architects microfrontend platforms and improves developer experience. His work focuses 
+        on building scalable, high-performance web applications that serve millions of users.
+      </p>
+      <h2>Technical Skills</h2>
+      <p>
+        Ivan Maier Gallardo's technical expertise includes React, TypeScript, JavaScript, Next.js, 
+        Node.js, Redux, and modern web development practices. He is passionate about performance 
+        optimization, developer experience, and creating maintainable codebases.
+      </p>
+      <h2>Contact Ivan Maier Gallardo</h2>
+      <p>
+        For professional inquiries, collaborations, or opportunities, you can reach Ivan Maier 
+        Gallardo at ivanmaiergallardo@gmail.com. Based in Buenos Aires, Argentina, Ivan Maier 
+        Gallardo is open to discussing frontend engineering opportunities and technical consulting.
+      </p>
+    </div>
+  )
+}
