+++ next.config.mjs
@@ -11,20 +11,15 @@ const nextConfig = {
     deviceSizes: [640, 750, 828, 1080, 1200, 1920, 2048, 3840],
     imageSizes: [16, 32, 48, 64, 96, 128, 256, 384],
   },
-  // Performance optimizations
   experimental: {
     optimizePackageImports: ['@radix-ui/react-icons'],
   },
   compiler: {
     removeConsole: process.env.NODE_ENV === 'production',
   },
-  // Enable compression
   compress: true,
-  // Optimize bundle
   swcMinify: true,
-  // Generate source maps for debugging
   productionBrowserSourceMaps: true,
-  // Reduce JavaScript bundle size
   modularizeImports: {
     '@radix-ui/react-icons': {
       transform: '@radix-ui/react-icons/dist/{{member}}',
